+++ tw-tkms-starter/src/main/java/com/transferwise/kafka/tkms/config/ITkmsDataSourceProvider.java
@@ -7,5 +7,5 @@ public interface ITkmsDataSourceProvider {
   /**
    * Provides the datasource for tw-tkms to use.
    */
-  DataSource getDataSource();
+  DataSource getDataSource(int shard);
 }