+++ tw-tkms-starter/src/main/java/com/transferwise/kafka/tkms/config/TkmsProperties.java
@@ -25,9 +25,11 @@ public class TkmsProperties {
   }
 
   /**
-   * Allows to set notification level or even block the startup, for different problems the library is detecting.
+   * Allows to set notification level or even block the startup/execution, for different problems the library is detecting.
    *
-   * <p>The set of keys is described with NotificationType class below.
+   * <p>The set of keys is described with {@link NotificationType} enum below.
+   * 
+   * <p>The level has to be one from {@link NotificationLevel} enum below. 
    */
   private Map<NotificationType, NotificationLevel> notificationLevels = new HashMap<>();
 
@@ -75,6 +77,7 @@ public class TkmsProperties {
    * <p>Make sure you have all those tables available and correctly configured - index statistics, auto vacuum and analyze settings.
    */
   @NotBlank
+  @ResolvedValue
   private String tableBaseName = "outgoing_message";
 
   /**
@@ -82,6 +85,7 @@ public class TkmsProperties {
    *
    * <p>Used to determine service-cluster wide lock names for Kafka proxies.
    */
+  @ResolvedValue
   private String groupId;
   /**
    * How many messages is Kafka proxy polling from a database at once.
@@ -143,6 +147,7 @@ public class TkmsProperties {
    *
    * <p>Be extra careful here by validating what is the corresponding value on the Kafka server side.
    */
+  @Positive
   private int maximumMessageBytes = 10485760;
 
   /**
@@ -166,6 +171,7 @@ public class TkmsProperties {
    */
   private List<String> topics = new ArrayList<>();
 
+  @Valid
   private EarliestVisibleMessages earliestVisibleMessages = new EarliestVisibleMessages();
 
   /**
@@ -180,10 +186,13 @@ public class TkmsProperties {
    */
   private Map<Integer, ShardProperties> shards = new HashMap<>();
 
+  @Valid
   private Compression compression = new Compression();
 
+  @Valid
   private Environment environment = new Environment();
 
+  @Valid
   private Monitoring monitoring = new Monitoring();
 
   /**
@@ -195,10 +204,27 @@ public class TkmsProperties {
   @NotNull
   private Internals internals = new Internals();
 
+  @Valid
+  @NotNull
+  private Mdc mdc = new Mdc();
+
+  @Data
+  @Accessors(chain = true)
+  public static class Mdc {
+
+    @NotBlank
+    @ResolvedValue
+    private String shardKey = "tkmsShard";
+    @NotBlank
+    @ResolvedValue
+    private String partitionKey = "tkmsPartition";
+  }
+  
   @Data
   @Accessors(chain = true)
   public static class ShardProperties {
 
+    @ResolvedValue
     private String tableBaseName;
     private DatabaseDialect databaseDialect;
     private Integer partitionsCount;
@@ -207,7 +233,9 @@ public class TkmsProperties {
     private Duration pauseTimeOnErrors;
     private Integer insertBatchSize;
     private boolean compressionOverridden;
+    @Valid
     private Compression compression = new Compression();
+    @Valid
     private EarliestVisibleMessages earliestVisibleMessages;
     private Boolean requireTransactionOnMessagesRegistering;
     private List<Integer> deleteBatchSizes;
@@ -324,6 +352,7 @@ public class TkmsProperties {
 
     private CompressionAlgorithm algorithm = CompressionAlgorithm.GZIP;
 
+    @Positive
     private Integer blockSize;
 
     /**
@@ -339,7 +368,7 @@ public class TkmsProperties {
   @Data
   @Accessors(chain = true)
   public static class Environment {
-
+    @ResolvedValue
     private String previousVersion;
   }
 
@@ -349,6 +378,7 @@ public class TkmsProperties {
 
     private boolean enabled = false;
 
+    @ResolvedValue
     private String tableName = "tw_tkms_earliest_visible_messages";
 
     private Duration lookBackPeriod = Duration.ofMinutes(5);
@@ -378,6 +408,7 @@ public class TkmsProperties {
   }
 
   public enum NotificationLevel {
+    IGNORE,
     INFO,
     WARN,
     ERROR,
@@ -396,6 +427,7 @@ public class TkmsProperties {
     TABLE_INDEX_STATS_CHECK_ERROR,
     TOO_MANY_DELETE_BATCHES,
     EARLIEST_MESSAGES_SYSTEM_DISABLED,
-    ENGINE_INDEPENDENT_STATS_NOT_ENABLED
+    ENGINE_INDEPENDENT_STATS_NOT_ENABLED,
+    NO_ACTIVE_TRANSACTION, EARLIEST_MESSAGES_SYSTEM_TABLE_NOT_EXISTING
   }
 }