+++ tw-tkms-starter/src/test/java/com/transferwise/kafka/tkms/dao/FaultInjectedTkmsDao.java
@@ -66,4 +66,14 @@ public class FaultInjectedTkmsDao implements ITkmsDao {
     return delegate.insertEarliestMessageId(shardPartition);
   }
 
+  @Override
+  public void validateDatabase() {
+    delegate.validateDatabase();
+  }
+
+  @Override
+  public void validateDatabase(int shard) {
+    delegate.validateDatabase(shard);
+  }
+
 }